requirements.txt
Flask==3.0.2
Pillow==10.2.0
pymongo==4.6.1
python-dotenv==1.0.1
transformers==4.34.1
watchdog==3.0.0